Move oodatamodel to the internal package
So, it's close to oonitemplates. This is yak shaving that I am
doing in preparation for #90.
