+++ internal/oonidatamodel/oonidatamodel.go
@@ -1,5 +1,8 @@
-// Package oodatamodel contains the OONI data model.
-package oodatamodel
+// Package oonidatamodel contains the OONI data model.
+//
+// The input of this package is data generated by netx and the
+// output is a format consistent with OONI specs.
+package oonidatamodel
 
 import (
 	"encoding/base64"
